Changes to get AROSTCP working on x86-64: incorporated new packet queue reassembly code, increased memory buffer size to take into account larger pointers, changed some types that always need to be 4-byte long from u_long to u_int32_t. Commited at once to allow easy reversal in case of problems with other architectures.
git-svn-id: https://svn.aros.org/svn/aros/trunk@28689 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/Networking/Stacks/AROSTCP/bsdsocket/api/gethostnamadr.c
@@ -241,7 +241,7 @@ D(bug("[AROSTCP](gethostnameadr.c) getanswer()\n"));
     type = _getshort(cp); 
     cp += sizeof(u_short);
     class = _getshort(cp);
-    cp += sizeof(u_short) + sizeof(u_long);
+    cp += sizeof(u_short) + sizeof(u_int32_t);
     n = _getshort(cp);
     cp += sizeof(u_short);
